Make "approve" hints require the tpu package to be built on all arches
Signed-off-by: Adam D. Barratt <adam@adam-barratt.org.uk>

Previously, an "approve" hint would allow a package to migrate from
testing-proposed-updates to testing even if packages were not available
for some architectures on which the package existed in testing.

The t-p-u package must now have built on the same architectures as the
existing package in testing; the previous behaviour can be achieved by
combining "approve" and "force" hints.