Properly remove failed tpu binNMUs from the result set
Signed-off-by: Adam D. Barratt <adam@adam-barratt.org.uk>

When undoing the addition to the result set of a binNMU from tpu, the package
name from the hint will end in ${arch}_tpu, not ${arch}, so we should check
for both cases.  At the same time, reverse the sense of the associated test
to make the logic more obvious.